requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
easyocr==1.7.0
editdistance==0.6.2
efficientnet==1.0.0
essential-generators==1.0
filelock==3.12.0
flatbuffers==23.5.26
fonttools==4.39.4
frozenlist==1.3.3
gast==0.4.0
gevent==22.10.2
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grequests==0.6.0
grpcio==1.54.2
h5py==3.8.0
idna==3.4
imageio==2.31.0
imgaug==0.4.0
jax==0.4.12
Jinja2==3.1.2
keras==2.12.0
Keras-Applications==1.0.8
keras-ocr==0.8.9
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
magic-filter==1.0.9
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
ninja==1.11.1
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pycparser==2.21
pyparsing==3.0.9
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tifffile==2023.4.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
validators==0.20.0
Werkzeug==2.3.6
wrapt==1.14.1
yarl==1.9.2
zope.event==4.6
zope.interface==6.0
